🚚 Move bulk endpoints to bulk router
